--- src/Shape.java
@@ -1,4 +1,0 @@
-
-public enum Shape {
-	shapeI , shapeJ , shapeL , shapeO , shapeS , shapeT , shapeZ
-}
