+++ src/fr/xnxa/tetris/Shape.java
@@ -1,0 +1,4 @@
+package fr.xnxa.tetris;
+public enum Shape {
+	shapeI , shapeJ , shapeL , shapeO , shapeS , shapeT , shapeZ
+}
