+++ src/module-info.java
@@ -1,0 +1,7 @@
+module Petris {
+	requires javafx.controls;
+    requires javafx.graphics;
+    requires javafx.fxml;
+	requires javafx.media;
+    opens fr.xnxa.tetris;
+}
